Showing Beneficiaries of member .

Approved-by: Russ Harding

+++ src/services/src/Demoulas.ProfitSharing.Common/Contracts/Request/BeneficiaryInquiry/BeneficiaryRequestDto.cs
@@ -9,6 +9,6 @@ namespace Demoulas.ProfitSharing.Common.Contracts.Request.BeneficiaryInquiry;
  public record BeneficiaryRequestDto: SortedPaginationRequestDto
 {
     public int? BadgeNumber { get; set; }
-    public short? PsnSuffix { get; set; }
+    public int? PsnSuffix { get; set; }
 
 }
